chore: move DownloadException and DownloadUtils to `remote` package
