+++ src/main/java/org/terasology/launcher/remote/DownloadException.java
@@ -1,7 +1,7 @@
-// Copyright 2021 The Terasology Foundation
+// Copyright 2023 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-package org.terasology.launcher.util;
+package org.terasology.launcher.remote;
 
 
 public final class DownloadException extends RuntimeException {
