+++ src/main/java/org/terasology/launcher/remote/DownloadUtils.java
@@ -1,7 +1,7 @@
-// Copyright 2021 The Terasology Foundation
+// Copyright 2023 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-package org.terasology.launcher.util;
+package org.terasology.launcher.remote;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
