+++ src/main/java/org/terasology/launcher/tasks/DownloadTask.java
@@ -8,7 +8,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.terasology.launcher.game.GameManager;
 import org.terasology.launcher.model.GameRelease;
-import org.terasology.launcher.util.DownloadException;
+import org.terasology.launcher.remote.DownloadException;
 
 import java.io.IOException;
